rename modmodals slice
