+++ src/extensionConsole/pages/mods/modals/convertToRecipeModal/ConvertToRecipeModal.tsx
@@ -17,8 +17,8 @@
 
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { selectModalsContext } from "@/extensionConsole/pages/mods/modals/blueprintModalsSelectors";
-import { blueprintModalsSlice } from "@/extensionConsole/pages/mods/modals/blueprintModalsSlice";
+import { selectModalsContext } from "@/extensionConsole/pages/mods/modals/modModalsSelectors";
+import { modModalsSlice } from "@/extensionConsole/pages/mods/modals/modModalsSlice";
 import { RequireScope } from "@/auth/RequireScope";
 import ModalLayout from "@/components/ModalLayout";
 import ConvertToRecipeModalBody from "./ConvertToRecipeModalBody";
@@ -26,7 +26,7 @@ import ConvertToRecipeModalBody from "./ConvertToRecipeModalBody";
 const ConvertToRecipeModal: React.FunctionComponent = () => {
   const dispatch = useDispatch();
   const closeModal = () => {
-    dispatch(blueprintModalsSlice.actions.closeModal());
+    dispatch(modModalsSlice.actions.closeModal());
   };
 
   const { showShareContext, showPublishContext } =
