+++ src/extensionConsole/pages/mods/modals/extensionLogsModal/ExtensionLogsModal.tsx
@@ -22,15 +22,15 @@ import { Modal } from "react-bootstrap";
 import { useDispatch, useSelector } from "react-redux";
 import LogCard from "@/components/logViewer/LogCard";
 import { logActions } from "@/components/logViewer/logSlice";
-import { blueprintModalsSlice } from "@/extensionConsole/pages/mods/modals/blueprintModalsSlice";
+import { modModalsSlice } from "@/extensionConsole/pages/mods/modals/modModalsSlice";
 import ModalLayout from "@/components/ModalLayout";
-import { selectShowLogsContext } from "@/extensionConsole/pages/mods/modals/blueprintModalsSelectors";
+import { selectShowLogsContext } from "@/extensionConsole/pages/mods/modals/modModalsSelectors";
 
 const ExtensionLogsModal: React.FunctionComponent = () => {
   const dispatch = useDispatch();
 
   const closeModal = () => {
-    dispatch(blueprintModalsSlice.actions.closeModal());
+    dispatch(modModalsSlice.actions.closeModal());
     dispatch(logActions.setContext(null));
   };
